updated parcelable wrappers to supported regeneration of wrapped objects.

+++ Core/src/org/droidplanner/core/drone/variables/GPS.java
@@ -58,7 +58,7 @@ public class GPS extends DroneVariable {
 		return fixType;
 	}
 
-	public void setGpsState(int fix, int satellites_visible, int eph) {
+	public void setGpsState(int fix, int satellites_visible, double eph) {
 		if (satCount != satellites_visible) {
 			satCount = satellites_visible;
 			gps_eph = (double) eph / 100; // convert from eph(cm) to gps_eph(m)
